Add optional support for memory checking tools
Sanitizers (address and leak) enable detecting some run-time memory
errors, such as accessing out-of-bounds memory locations (in both
heap and stack), and memory leaks. Unfortunately, gcc doesn't seem
to support memory sanitizer at the moment, so uninitilized memory
access is not with these tools. Sanitizers can be enabled
by giving -DENABLE_SANITIZER=ON parameter to cmake, for example:

$ cmake -DENABLE_SANITIZERS=ON .

..and then running make as usual. As sanitizers are compiler
feature, they will be included in all binaries built with this
parameter on.

Valgrind is another tool for detecting memory errors such as
accessing unitialized memory, accessing out-of-bound memory in
heap, and memory leaks. CMake will now check if valgrind program
is found, and if it is (and sanitizers are not enabled), the
spindump_test test will be executed in valgrind.

If any errors are found by sanitizers or valgrind during tests,
it will fail the make command.

This commit is initial work on issue #87.